[정연재] 2차 코드정리 및 주석
