+++ src/components/templates/PostDetail.jsx
@@ -1,35 +1,23 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 
-// Swiper
-import { Swiper, SwiperSlide } from 'swiper/react';
-import SwiperCore, { Pagination } from 'swiper/core';
-import 'swiper/swiper.min.css';
-import 'swiper/components/pagination/pagination.min.css';
-
 // Component
 import Header from '../modules/layout/Header';
-import MainComment from '../../components/modules/Post/MainComment';
-import { Grid, Image, Tag } from '../element/index';
-import { ReactComponent as HeartIcon } from '../../img/Icon/icon_heart_no.svg';
-import { ReactComponent as OnHeartIcon } from '../../img/Icon/icon_heart_in.svg';
-import { ReactComponent as CommentIcon } from '../../img/Icon/chat_bubble.svg';
+import PostCard from '../modules/Post/PostCard';
+import PostModal from '../modules/Post/PostModal';
+import AddComment from '../modules/Post/AddComment';
+import MainComment from '../modules/Post/MainComment';
+import { Grid } from '../element/index';
 
 // Redux
 import { useDispatch, useSelector } from 'react-redux';
 import { actionCreators as postActions } from '../../redux/modules/post';
 import { actionCreators as userAction } from '../../redux/modules/user';
-import PostModal from '../modules/Post/PostModal';
-import AddComment from '../modules/Post/AddComment';
-
-// Js
-import icon_location from '../../img/Icon/icon_location.svg';
 
 const PostDetail = props => {
   const dispatch = useDispatch();
-  SwiperCore.use([Pagination]);
 
-  // path
+  // 포스트 아이디를 가져옵니다.
   const pathName = props.location.pathname;
   const boardId = pathName.split('/');
 
@@ -39,34 +27,34 @@ const PostDetail = props => {
 
   // 예외처리용 정보
   const detailInfo = useSelector(state => state.post.detail);
-  const mbti = detailInfo.mbti;
-  const imgList = detailInfo.imageList;
   const comment = detailInfo.commentList;
-  const time = detailInfo.createdAt ? detailInfo.createdAt.split(' ')[0].split('-') : 0;
 
+  // 디테일 페이지 정보를 받아옵니다.
   React.useEffect(() => {
     dispatch(postActions.getDetailDB(boardId[2]));
   }, []);
 
+  // ???
   React.useEffect(() => {
     dispatch(userAction.getMyPostBoxDB());
   }, []);
 
+  // 모달창 관리
   const [modal, setModal] = useState(false);
-  const [commentState, setCommentState] = useState(false);
-
-  const [commenttId, setCommenttId] = useState();
-
-  const onCommentModal = () => {
-    setModal(true);
-    setCommentState(true);
-  };
-
   const outModal = () => {
     setModal(false);
     setCommentState(false);
   };
 
+  // 코멘트 관리
+  const [commentState, setCommentState] = useState(false);
+  const [commenttId, setCommenttId] = useState();
+  const onCommentModal = () => {
+    setModal(true);
+    setCommentState(true);
+  };
+
+  // 좋아요 이벤트
   const addlike = () => {
     dispatch(postActions.addLikeDB(boardId[2]));
   };
@@ -84,68 +72,10 @@ const PostDetail = props => {
         게시글 상세보기
       </Header>
 
-      <Grid>
-        <Grid borderTop="1px solid #E8E8E8" row wrap="nowrap" padding="18px 30px" gap="10px">
-          <Grid width="auto">
-            <Image src={detailInfo.profileImage} photoRound width="50px" margin="0" />
-          </Grid>
-          <Grid gap="10px" justify="center">
-            <Grid row gap="13px">
-              <NameText>{detailInfo.nickname}</NameText>
-              <Tag mbti={mbti ? mbti : 'INFJ'} _type="black" icon>
-                {detailInfo.mbti}
-              </Tag>
-            </Grid>
-            <TimeText>
-              {Number(time[1])}월 {time[2]}일
-            </TimeText>
-          </Grid>
-          <Grid gap="9px" justify="center" align="flex-end" width="60%">
-            <LocalText>
-              <img alt="마커" src={icon_location} />
-              서울특별시 {detailInfo.location}
-            </LocalText>
-            <Tag mbti={mbti ? mbti : 'INFJ'} padding="4px 10px" size="12px">
-              {detailInfo.tag}
-            </Tag>
-          </Grid>
-        </Grid>
+      {/* 포스트 창 */}
+      {detailInfo.length !== 0 ? <PostCard addlike={addlike} info={detailInfo} /> : null}
 
-        <Swiper
-          className="swiper-container"
-          style={{ width: '100%', position: 'relative' }}
-          pagination={{
-            type: 'fraction',
-          }}
-        >
-          {imgList
-            ? imgList[0].imageLink === ''
-              ? null
-              : imgList.map((list, idx) => {
-                  return (
-                    <SwiperSlide key={list.imageId}>
-                      <Image src={list.imageLink} />
-                    </SwiperSlide>
-                  );
-                })
-            : null}
-          <PaginationBakc />
-        </Swiper>
-
-        <Grid padding="17px 44px 16px 33px">
-          <Text>{detailInfo.content}</Text>
-          <Grid row padding="16px 0px 0px 0px" align="center">
-            {detailInfo.likeStatus ? (
-              <OnHeartIcon style={{ cursor: 'pointer' }} onClick={addlike} />
-            ) : (
-              <HeartIcon style={{ cursor: 'pointer' }} onClick={addlike} />
-            )}
-            <Count>{detailInfo.likesCount}</Count>
-            <CommentIcon style={{ margin: '0px 0px 0px 14px' }} />
-            <Count>{comment ? comment.length : null}</Count>
-          </Grid>
-        </Grid>
-      </Grid>
+      {/* 댓글창 및 댓글 등록 */}
       <Grid padding="20px 30px" gap="12px" borderTop="1px solid #E8E8E8">
         {comment
           ? comment.map((x, idx) => {
@@ -163,6 +93,8 @@ const PostDetail = props => {
           : null}
       </Grid>
       <AddComment boardId={boardId[2]}></AddComment>
+
+      {/* 모달창 */}
       {modal ? (
         <PostModal
           commenttId={commenttId}
@@ -174,51 +106,6 @@ const PostDetail = props => {
     </DetailBox>
   );
 };
-
-const NameText = styled.p`
-  font-weight: 700;
-  font-size: ${props => props.theme.fontSizes.xl};
-`;
-
-const TimeText = styled.p`
-  font-weight: 400;
-  font-size: ${props => props.theme.fontSizes.base};
-  color: #9b9b9b;
-`;
-
-const LocalText = styled.p`
-  font-weight: 500;
-  font-size: ${props => props.theme.fontSizes.extraSmall};
-  color: #9b9b9b;
-  display: flex;
-  align-items: center;
-`;
-
-const Count = styled.p`
-  font-weight: 500;
-  font-size: ${props => props.theme.fontSizes.small};
-  margin: 0px 0px 0px 4px;
-`;
-
-const Text = styled.p`
-  display: block;
-  word-break: break-all;
-  line-height: 1.5;
-  font-weight: 500;
-  font-size: ${props => props.theme.fontSizes.small};
-`;
-const PaginationBakc = styled.div`
-  width: 80px;
-  height: 25px;
-  background: #fff;
-  position: absolute;
-  bottom: 6px;
-  z-index: 1;
-  border-radius: 50px;
-  opacity: 0.5;
-  left: 50%;
-  margin-left: -40px;
-`;
 
 const DetailBox = styled.div`
   height: calc(100% + 105px);
